Move ast_string to pool
