+++ src/ast_method.h
@@ -6,6 +6,7 @@
 #include "ast/declarations.h"
 #include "ast/statements.h"
 #include "ast/external_definitions.h"
+#include "ast/pool.h"
 
 void ast_initialize_pool(void);
 void ast_finalize_pool(void);
